# Fix possible memory confusion in unsafe slice cast