@Fenny 2020-05-30
Hi @jlauinger, thank you for your detailed explanation!

We use the same method as [Fasthttp](https://github.com/valyala/fasthttp/blob/master/bytesconv.go#L332) to convert strings and bytes without memory allocation, but requires you to never change the original byte array, but that's a reasonable contract in many cases. To persist values after returning from the handler, you can use the build-in [Immutable](https://docs.gofiber.io/application#settings) setting in Fiber.

```go
app := fiber.New(&fiber.Settings{
    Immutable: true,
})
```

Having said the above, the garbage collector will indeed clean-up the underlying data if you return from the `fiber.handler` and this is expected behavior. This won't happen between conversion when you are inside the handler.

I also run some tests to see if it would impact performance
```go
// current
Benchmark_GetBytes/fiber-4              483916431                2.43 ns/op            0 B/op          0 allocs/op
Benchmark_GetBytes/fiber-4              491261481                2.42 ns/op            0 B/op          0 allocs/op

// new
Benchmark_GetBytes/fiber-4              172217029                6.94 ns/op            0 B/op          0 allocs/op
Benchmark_GetBytes/fiber-4              171653289                6.90 ns/op            0 B/op          0 allocs/op
```