@jlauinger 2020-05-30
Hi @Fenny, thanks for your comment and looking into this! Right, so there is two different concerns with the string/slice casting:

 1. When converting a string to a slice in-place, the resulting slice might be effectively read-only (in fact that is the mimimal guarantee Go makes) because the string was immutable in the first place and might live within a read-only memory page. This is something developers need to keep in mind when dealing with the result value of such casts, but it is not the concern of this PR.

 2. The garbage collector might clean up the underlying array of the source (e.g. string) just after there are no more references to it (because both the source variable and the `reflect.StringHeader`, or be it `reflect.StringHeader` depending on the cast direction, resulting from it are not used anymore) and just before the destination slice (or string) is created. In that exact moment, the only reference to the underlying data is in the destination `reflect.SliceHeader`, and when it is not created from a cast but by a composite literal then Go will not correctly treat it as an actual reference.

As far as i know, Go treats actual pointers, `unsafe.Pointer` values, and `uintptr` values *within a slice/string header that was cast from an actual slice/string* as references. Importantly, it does *not* treat plain `uintptr` values as a reference.

Garbage collection might indeed happen within the conversion function, albeit pretty unlikely. I crafted a proof of concept for this that is going to be published on DEV in a couple of days, if you are interested I'd be happy to send it to you.

Fasthttp, in the code you linked, actually uses a just slightly different variation of your code: they create their `reflect.SliceHeader` by a cast instead of a composite literal. That makes all the difference here.

However, I certainly didn't expect this change to have such a dramatic impact on performance. I honestly don't know if it's worth it either...