@jlauinger 2020-06-03
In fact another PR in a different project pointed out that the slice doesn't have to be initialized, only declared, so I updated this PR to skip the `make` call. This should drop the performance problems while still keeping the safer cast.