@jlauinger 2020-06-03
The only difference is the order in which `reflect.StringHeader` and `reflect.SliceHeader` are created by cast, right? It's pretty much the same code isn't it?

What do the four lines in the benchmark result mean? Is it the same benchmark run four times? So your results would show that the two variants are in fact equal, with some deviation of +/- 10ns/op on both?