@Fenny 2020-06-03
Yeah, they are both equal regarding the performance. But I meant the following difference, if it would matter.
```go
sh := *(*reflect.StringHeader)(unsafe.Pointer(&s))
// vs 
sh := (*reflect.StringHeader)(unsafe.Pointer(&s))
```