@jlauinger 2020-06-03
Ohhh, right! Didn't even spot it! I think it just doesn't matter. You can't dereference the target (`bh`) because then the cast will simply not work, but I think whether you use a direct `StringHeader` or a pointer to it as the source is irrelevant both with respect to security and apparently performance.

The only argument in favor of my variant I can think of is symmetry. Like, think about what happens if someone copies the code but for a cast in the other direction (from `[]byte` to `string), then they would have to adjust the pointers, too. But it's really a very minor thing and boils down to pretty much nothing except your personal preference.