@Fenny 2020-06-03
Yeah it since it's already a reference the first seems unnecessary so I prefer the second. I made a co-pr and will tag it in the next version. Thank you for this discussion @jlauinger 👍